+++ .github/workflows/build.yml
@@ -23,7 +23,7 @@ jobs:
         run: pip install .[develop]
       - name: Run pylint
         run: |
-          pylint src/gnn_reco tests/*.py --exit-zero | tee results_pylint.txt
+          pylint src/graphnet tests/*.py --exit-zero | tee results_pylint.txt
       - name: Generate badge
         run: grep rated < results_pylint.txt | sed 's/.* at //g;s/\/.*//g' | xargs -I {} anybadge -f pylint.svg -v {} pylint
       - name: Archive pylint results and badge
@@ -32,25 +32,25 @@ jobs:
           name: pylint
           path: |
             pylint.svg
-          
+
   tests:
     runs-on: ubuntu-latest
     strategy:
       matrix:
         python-version: [3.6, 3.7, 3.8, 3.9]
-      
+
     steps:
       - uses: actions/checkout@v2  # Checks-out your repository under $GITHUB_WORKSPACE, so your job can access it
       - name: Set up Python ${{ matrix.python-version }}
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
-      - name: Install required packages 
+      - name: Install required packages
         run: pip install .[develop]
       - name: Run unit tests and generate coverage report
         run: |
           set -o pipefail  # To propagate exit code from pytest
-          coverage run --source=gnn_reco -m pytest tests/ | tee results_pytest.txt
+          coverage run --source=graphnet -m pytest tests/ | tee results_pytest.txt
           coverage report -m | tee results_coverage.txt
       - name: Generate unit test/build and coverage badges
         run: |
@@ -59,9 +59,9 @@ jobs:
         uses: actions/upload-artifact@v2
         with:
           name: coverage
-          path: | 
+          path: |
             coverage.svg
-  
+
   badges:
     if: ${{ github.event_name == 'push' }}
     needs: [ linting, tests ]
@@ -89,7 +89,7 @@ jobs:
       - name: Stage badges
         run: git add -f misc/badges/
       - name: Commit and push changes
-        run: git commit -m "Auto-updating badges" > /dev/null && git push origin ${GITHUB_REF##*/} || echo "Nothing to push"  
+        run: git commit -m "Auto-updating badges" > /dev/null && git push origin ${GITHUB_REF##*/} || echo "Nothing to push"
 
   docs:
     if: ${{ github.event_name == 'push' && github.ref == 'refs/heads/main' }}
