+++ README.md
@@ -1,7 +1,8 @@
-# gnn-reco
-Development of a GNN based IceCube / DeepCore / Upgrade reconstruction
+# GraphNeT
 
-![build](https://github.com/icecube/gnn-reco/actions/workflows/build.yml/badge.svg)
+Graph neural networks for neutrino telescope event reconstruction.
+
+![build](https://github.com/icecube/graphnet/actions/workflows/build.yml/badge.svg)
 ![pylint](./misc/badges/pylint.svg)
 ![coverage](./misc/badges/coverage.svg)
 [![License](https://img.shields.io/badge/License-Apache%202.0-blue.svg)](https://opensource.org/licenses/Apache-2.0)
