+++ docker/gnn-benchmarking/dockerfile
@@ -11,7 +11,7 @@ RUN pip3 install torch==${TORCH}+${CUDA} -f https://download.pytorch.org/whl/cpu
     pip3 install torch-spline-conv -f https://data.pyg.org/whl/torch-${TORCH}+${CUDA}.html  && \
     pip3 install torch-geometric==${PYG}
 
-RUN pip3 install git+https://github.com/icecube/gnn-reco.git
+RUN pip3 install git+https://github.com/icecube/graphnet.git
 
 WORKDIR /root
