+++ src/graphnet/components/utils.py
