+++ src/graphnet/data/constants.py
