+++ src/graphnet/legacy/callbacks.py
