+++ src/graphnet/models/detector/__init__.py
