+++ src/graphnet/models/gnn/convnet.py
@@ -13,7 +13,7 @@ import torch.nn.functional as F
 from torch_geometric.nn import TAGConv, global_add_pool, global_max_pool
 from torch_geometric.data import Data
 
-from gnn_reco.models.gnn.gnn import GNN
+from graphnet.models.gnn.gnn import GNN
 
 
 class ConvNet(GNN):
