+++ src/graphnet/models/task/task.py
@@ -11,7 +11,7 @@ from torch import Tensor
 from torch.nn import Linear
 from torch_geometric.data import Data
 
-from gnn_reco.components.loss_functions import LossFunction
+from graphnet.components.loss_functions import LossFunction
 
 
 class Task(LightningModule):
