+++ tests/test_data_utils.py
@@ -1,6 +1,6 @@
 """Unit tests for I3Extractor class."""
 
-from gnn_reco.data.utils import is_i3_file, has_extension
+from graphnet.data.utils import is_i3_file, has_extension
 
 
 def test_is_i3_file():
